refactor(test): rename regression_v2 to prediction_regression_gate + accuracy_tracking
Split by signal: tools/prediction_regression_gate/ (old-vs-new snapshot
diff + tier2, the PR gate) and tools/accuracy_tracking/ (silicon anchors,
prediction vs measured, the nightly workload). 'accuracy' is reserved for
comparison against silicon; the gate compares prediction vs prediction.

Mechanical rename: workflow file, design doc, test files, imports, and
artifact/dir names follow; no behavior change (unit tests + e2e smoke
pass unchanged).
